LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-is-still-set-when-coercing-beforeunload-result-expected.txt



PASS window.event is still set when 'beforeunload' result is coerced to string
